# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Restaurant API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Авторизация",
                "operationId": "c373bc4cffa4a56fc64e3979b8ee7b96",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "email": {
                                                "type": "string",
                                                "example": "email@gmail.com"
                                            },
                                            "password": {
                                                "type": "string",
                                                "example": "123456789"
                                            },
                                            "pin_code": {
                                                "type": "string",
                                                "example": "1234"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Пользователь авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "type": "string",
                                                    "example": "1|tW9TJIbIxVrKnjeJ7Pfg2gDeQb5lOS7KuucGtEkLee9da4cd"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Неправильный логин и/или пароль",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Error": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "Message": {
                                            "type": "string",
                                            "example": "Неверный пароль или пин-код"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Error": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "Message": {
                                            "type": "string",
                                            "example": "email has been requred"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Error": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "Message": {
                                            "type": "string",
                                            "example": "Error text"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Выход",
                "operationId": "5ff2dc8b48bff52908d3821a0ecc9b3e",
                "responses": {
                    "200": {
                        "description": "Выход произведен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Message": {
                                            "type": "string",
                                            "example": "Выход произведён"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Error": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "Message": {
                                            "type": "string",
                                            "example": "Error text"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/logout_all": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Выход со всех устройств",
                "operationId": "7a8410f30e379e0c7c88c4cdbb46d50f",
                "responses": {
                    "200": {
                        "description": "Выход со всех устройсв произведен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Message": {
                                            "type": "string",
                                            "example": "Произведён выход со всех устройств!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Error": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "Message": {
                                            "type": "string",
                                            "example": "Error text"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/forgot-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Сброс пароля",
                "operationId": "5329cdf94f8be3890847c6d96d32f8bd",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "email": {
                                                "type": "string",
                                                "example": "email@gmail.com"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "пароль сброшен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "We send mail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка на клиенте",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "email": {
                                            "type": "string",
                                            "example": "email error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Error": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "Message": {
                                            "type": "string",
                                            "example": "email has been requred"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Error": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "Message": {
                                            "type": "string",
                                            "example": "Error text"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/reset-password/{token}": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Восстановление пароля",
                "operationId": "37641ed6bd7bf7b6c3e114b9d05fd512",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "description": "Токен",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "c1bb927ef7fdd9c1efda0d2ee650f8887ba5c1fcd4c573ab51176316a07baba8"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "email@gmail.com"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "token": {
                                                "type": "string",
                                                "example": "c1bb927ef7fdd9c1efda0d2ee650f8887ba5c1fcd4c573ab51176316a07baba8"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "email@gmail.com"
                                            },
                                            "password": {
                                                "type": "string",
                                                "example": "123456789"
                                            },
                                            "password_confirmation": {
                                                "type": "string",
                                                "example": "123456789"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Пароль восстановлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Password has been restored"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка на клиенте",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "email": {
                                            "type": "string",
                                            "example": "email error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Error": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "Message": {
                                            "type": "string",
                                            "example": "email has been requred"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Error": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "Message": {
                                            "type": "string",
                                            "example": "Error text"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}